Add test to make sure we compare with intl
Using only the "type" arguments to the intl formatter,
we can mimic the special localized formats (that are not a part of DateTime formatting).
Only the "llll" format does not work that way, and we need a pattern to test that out.

Links at the top of the file can be used to verify.